Moved `EngineControls` to `engine_controls.rs`
